refactor(registry-lint): avoid relying on shell scripts to run node scripts
Instead of parsing yaml files in shell and passing via stdin as a string we instead do the parsing directly in Node via
the yaml dependency. This requires bundling the Node script to access dependencies when executing the script on GitHub
runners.

Also fixes the implementation of checking invalid languages by only considering languages included in files that are
changed, not the entire registry (i.e. PRs won't fail if some other packages contain invalid languages).