refactor(tests): use vim.fs instead of vim.fn.fnamemodify
While it's probably safe to call vim.fn here, we want to avoid doing so as much as possible because this can only be
done when executing in the main event loop. mason.nvim, and its user callbacks, executes a large chunk of its code
outside of the main event loop due to frequent async interactions with libuv.

Oh, also better readability :P